pyproject.toml
[project]
name = "2021-spotify-music-adder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "spotipy==2.23.0",
    "pandas==2.2.1",
    "python-dotenv==1.0.1",
    "tidalapi>=0.8.3",
    "ipykernel>=6.29.5",
]
